move base truchet classes up for use by both imagefilter and turtlegenerator
