+++ src/main/java/com/marginallyclever/makelangelo/makeart/imageconverter/Converter_TruchetFromImage.java
@@ -3,7 +3,7 @@ package com.marginallyclever.makelangelo.makeart.imageconverter;
 import com.marginallyclever.makelangelo.Translator;
 import com.marginallyclever.makelangelo.makeart.TransformedImage;
 import com.marginallyclever.makelangelo.makeart.imagefilter.FilterDesaturate;
-import com.marginallyclever.makelangelo.makeart.turtlegenerator.truchet.TruchetDiagonal;
+import com.marginallyclever.makelangelo.makeart.truchet.TruchetDiagonal;
 import com.marginallyclever.makelangelo.paper.Paper;
 import com.marginallyclever.makelangelo.select.SelectSlider;
 import com.marginallyclever.makelangelo.turtle.Turtle;
