+++ src/main/java/com/marginallyclever/makelangelo/makeart/truchet/TruchetDiagonal.java
@@ -1,13 +1,13 @@
-package com.marginallyclever.makelangelo.makeart.turtlegenerator.truchet;
+package com.marginallyclever.makelangelo.makeart.truchet;
 
 import com.marginallyclever.makelangelo.turtle.Turtle;
 
-public class TruchetOrthogonal implements TruchetTileGenerator {
+public class TruchetDiagonal implements TruchetTile {
     private final Turtle turtle;
     private final double spaceBetweenLines;
     private final double tileSize;
 
-    public TruchetOrthogonal(Turtle turtle, int spaceBetweenLines, int linesPerTileCount) {
+    public TruchetDiagonal(Turtle turtle,double spaceBetweenLines,double linesPerTileCount) {
         super();
         this.turtle=turtle;
         this.spaceBetweenLines = spaceBetweenLines;
@@ -20,29 +20,31 @@ public class TruchetOrthogonal implements TruchetTileGenerator {
         else                     tileB(x,y);
     }
 
-    // horizontal lines
+    // style=/
     public void tileA(double x0,double y0) {
+        double x1=x0+tileSize;
         double y1=y0+tileSize;
 
-        for(double x=0;x<=tileSize;x += spaceBetweenLines) {
-            interTile(
-                    x0+x,y0,
-                    x0+x,y1);
+        for(double x=spaceBetweenLines;x<tileSize;x += spaceBetweenLines) {
+            interTile(x0+x,y0,x0,y0+x);
+            interTile(x0+x,y1,x1,y0+x);
         }
+        interTile(x0,y1,x1,y0);
     }
 
-    // vertical lines
+    // style=\
     public void tileB(double x0,double y0) {
         double x1=x0+tileSize;
         double y1=y0+tileSize;
 
-        for(double y=0;y<=tileSize;y += spaceBetweenLines) {
-            interTile(
-                    x0,y0+y,
-                    x1,y0+y);
+        for(double x=spaceBetweenLines;x<tileSize;x += spaceBetweenLines) {
+            interTile(x0+x,y0,x1,y1-x);
+            interTile(x0+x,y1,x0,y1-x);
         }
+        interTile(x0,y0,x1,y1);
     }
 
+    // Interpolate from (x0,y0) to (x1,y1) in steps of length iterSize.
     private void interTile(double x0,double y0,double x1,double y1) {
         turtle.penUp();
         turtle.moveTo(x0,y0);
