+++ src/main/java/com/marginallyclever/makelangelo/makeart/truchet/TruchetTile.java
@@ -1,0 +1,8 @@
+package com.marginallyclever.makelangelo.makeart.truchet;
+
+/**
+ * A single Truchet tile.
+ */
+public interface TruchetTile {
+    void drawTile(double x,double y);
+}
