--- src/main/java/com/marginallyclever/makelangelo/makeart/turtlegenerator/truchet/TruchetTileGenerator.java
@@ -1,6 +1,0 @@
-package com.marginallyclever.makelangelo.makeart.turtlegenerator.truchet;
-
-public interface TruchetTileGenerator {
-    void drawTile(double x,double y);
-
-}
